Renames audio player, moves it out of microphone dir with async util
